change float -> double

+++ include/vector_lib.h
@@ -10,21 +10,21 @@ public:
   Vector();
   Vector(Point p);
   Vector(Point a, Point b);
-  Vector(float x, float y, float z);
+  Vector(double x, double y, double z);
 
-  float x() const { return p.x; };
-  float y() const { return p.y; };
-  float z() const { return p.z; };
-  float mod() const;
-  float mod_square() const;
+  double x() const { return p.x; };
+  double y() const { return p.y; };
+  double z() const { return p.z; };
+  double mod() const;
+  double mod_square() const;
 
   Vector unit() const;
-  Vector scale(float factor) const;
+  Vector scale(double factor) const;
   Vector direction_cossines() const;
   Vector opposite() const;
-  Vector rotate_around_x(float angle) const;
-  Vector rotate_around_y(float angle) const;
-  Vector rotate_around_z(float angle) const;
+  Vector rotate_around_x(double angle) const;
+  Vector rotate_around_y(double angle) const;
+  Vector rotate_around_z(double angle) const;
   Vector reflect_across_plane(const Vector &normal) const;
 
   bool is_unit() const;
@@ -32,14 +32,14 @@ public:
 
   Vector operator+(const Vector &) const;
   Vector operator-(const Vector &) const;
-  Vector operator*(float) const;
-  float operator*(const Vector &) const;
-  Vector operator/(float) const;
+  Vector operator*(double) const;
+  double operator*(const Vector &) const;
+  Vector operator/(double) const;
   bool operator==(const Vector &) const;
   bool operator!=(const Vector &) const;
 
-  static float angle(const Vector &, const Vector &);
-  static float mix_product(const Vector &, const Vector &, const Vector &);
+  static double angle(const Vector &, const Vector &);
+  static double mix_product(const Vector &, const Vector &, const Vector &);
   static Vector cross_product(const Vector &, const Vector &);
   static Vector projection(const Vector &, const Vector &);
